+++ Server/Controllers/TomSoftwareController.cs
@@ -1,47 +1,56 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Storage;
 using Software.Service;
+using TomSoftware.Service;
+using WorkShopContext;
 
 namespace Server.Controllers;
 
 [ApiController]
-[Route("api/software")]
-public class SoftwareController : ControllerBase
+[Route("api/TomSoftware")]
+public class TomSoftwareController : ControllerBase
 {
-    SoftwareService _softwareService;
-    public SoftwareController(SoftwareService softwareService)
-    {
-        _softwareService = softwareService;
-    }
+    private readonly IDbContextFactory<WorkShopDBContext> _contextFactory;
+
+    public TomSoftwareController(IDbContextFactory<WorkShopDBContext> contextFactory)
+        => _contextFactory = contextFactory;
+
     [HttpPost("Create")]
-    public IResult Create([FromHeader(Name = "Name")] string name, [FromHeader(Name = "Price")] double price)
+    public IResult Create()
     {
-        return Results.Ok(_softwareService.Create(name, price));
-    }
-
-    [HttpPost("FindByID")]
-    public IResult FindByID([FromHeader(Name = "Id")] int id)
-    {
-        return Results.Ok(_softwareService.FindById(id));
-    }
-
-    [HttpGet("FindAll")]
-    public IResult FindAll()
-    {
-        return Results.Ok(_softwareService.FindAll());
-    }
-
-    [HttpDelete("Delete")]
-    public IResult DelteByID([FromHeader(Name = "Id")] int id)
-    {
-        _softwareService.DeleteById(id);
+        var _db = _contextFactory.CreateDbContext();
+        _db.Database.EnsureCreated();
+        var databaseCreator = _db.GetService<IRelationalDatabaseCreator>();
+        databaseCreator.CreateTables();
         return Results.Ok();
     }
 
-    [HttpPatch("Update")]
-    public IResult Update([FromHeader(Name = "Id")] int id, [FromHeader(Name = "Name")] string name, [FromHeader(Name = "Price")] double price)
-    {
-        _softwareService.Update(new() { Id = id, name = name, price = price });
-        return Results.Ok();
-    }
+    //[HttpPost("FindByID")]
+    //public IResult FindByID([FromHeader(Name = "Id")] int id)
+    //{
+    //    return Results.Ok(_softwareService.FindById(id));
+    //}
+
+    //[HttpGet("FindAll")]
+    //public IResult FindAll()
+    //{
+    //    return Results.Ok(_softwareService.FindAll());
+    //}
+
+    //[HttpDelete("Delete")]
+    //public IResult DelteByID([FromHeader(Name = "Id")] int id)
+    //{
+    //    _softwareService.DeleteById(id);
+    //    return Results.Ok();
+    //}
+
+    //[HttpPatch("Update")]
+    //public IResult Update([FromHeader(Name = "Id")] int id, [FromHeader(Name = "Name")] string name, [FromHeader(Name = "Price")] double price)
+    //{
+    //    _softwareService.Update(new() { Id = id, name = name, price = price });
+    //    return Results.Ok();
+    //}
 
 }
